djangogramm/requirements.txt
asgiref==3.7.2
boto3==1.34.144
botocore==1.34.144
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==4.2.10
django-allauth==64.2.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-resized==1.0.2
django-storages==1.14.4
fido2==1.1.3
idna==3.7
jmespath==1.0.1
oauthlib==3.2.2
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.2
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.9.0
urllib3==1.26.19
